examples/canonical-auth/python/pyproject.toml
[project]
name = "canonical-auth-python"
version = "0.1.0"
description = "TenantCore IAM Python Canonical Example"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[tool.uv]
managed = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
